+++ examples/storage_volume_attachments.py
@@ -27,19 +27,19 @@ from hpOneView.exceptions import HPOneViewException
 from config_loader import try_load_from_file
 
 config = {
-    "ip": "172.16.102.59",
+    "ip": "<oneview_ip>",
     "credentials": {
-        "userName": "administrator",
-        "password": ""
+        "userName": "<username>",
+        "password": "<password>"
     }
 }
 
 # To run all parts of this example, a server profile uri, volume uri, volume attachment id and
 # path id must be defined.
-serverProfileUri = ""
-storageVolumeUri = ""
-attachment_id = ""
-path_id = ""
+serverProfileUri = None
+storageVolumeUri = None
+attachment_id = None
+path_id = None
 
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
@@ -47,23 +47,25 @@ config = try_load_from_file(config)
 oneview_client = OneViewClient(config)
 
 # Get all volume attachments
-print("Get all volume attachments")
+print("\nGet all volume attachments")
 volume_attachments = oneview_client.storage_volume_attachments.get_all()
 for attachment in volume_attachments:
-    print("   'hostName': '{hostName}' at uri: '{uri}'".format(**attachment))
+    print('\n#### Storage Volume Attachment info:')
+    pprint(attachment)
 
 # Get all storage volume attachments filtering by storage volume URI
 try:
-    print("Get all storage volume attachments filtering by storage volume URI")
+    print("\nGet all storage volume attachments filtering by storage volume URI")
     filter = "storageVolumeUri='{}'".format(storageVolumeUri)
     volume_attachments_filtered = oneview_client.storage_volume_attachments.get_all(filter=filter)
     for attachment in volume_attachments_filtered:
-        print("   'hostName': '{hostName}' at uri: '{uri}'".format(**attachment))
+        print('\n#### Storage Volume Attachment info:')
+        pprint(attachment)
 except HPOneViewException as e:
     print(e.msg)
 
 # Get the list of extra unmanaged storage volumes
-print("Get the list of extra unmanaged storage volumes")
+print("\nGet the list of extra unmanaged storage volumes")
 unmanaged_storage_volumes = oneview_client.storage_volume_attachments.get_extra_unmanaged_storage_volumes()
 pprint(unmanaged_storage_volumes)
 
@@ -73,7 +75,7 @@ try:
         "type": "ExtraUnmanagedStorageVolumes",
         "resourceUri": serverProfileUri
     }
-    print("Removes extra presentations from a specified server profile at uri: '{}".format(serverProfileUri))
+    print("\nRemoves extra presentations from a specified server profile at uri: '{}".format(serverProfileUri))
     oneview_client.storage_volume_attachments.remove_extra_presentations(info)
     print("   Done.")
 except HPOneViewException as e:
@@ -81,35 +83,39 @@ except HPOneViewException as e:
 
 # Get storage volume attachment by id
 try:
-    print("Get storage volume attachment by id: '{}'".format(attachment_id))
+    print("\nGet storage volume attachment by id: '{}'".format(attachment_id))
     volume_attachment_byid = oneview_client.storage_volume_attachments.get(attachment_id)
-    print("   Found '{hostName}' at uri: {uri}".format(**volume_attachment_byid))
+    print('\n#### Storage Volume Attachment info:')
+    pprint(attachment)
 except HPOneViewException as e:
     print(e.msg)
 
 if volume_attachments:
     # Get storage volume attachment by uri
-    print("Get storage volume attachment by uri: '{uri}'".format(**volume_attachments[0]))
+    print("\nGet storage volume attachment by uri: '{uri}'".format(**volume_attachments[0]))
     volume_attachment_byid = oneview_client.storage_volume_attachments.get(volume_attachments[0]['uri'])
-    print("   Found '{hostName}' at uri: {uri}".format(**volume_attachment_byid))
+    print('\n#### Storage Volume Attachment info:')
+    pprint(attachment)
 
-    # Get all volume attachment paths
-    print("Get all volume attachment paths for volume attachment at uri: {uri}".format(**volume_attachment_byid))
-    paths = oneview_client.storage_volume_attachments.get_paths(volume_attachment_byid['uri'])
-    for path in paths:
-        print("   Found path at uri: {uri}".format(**path))
+    if oneview_client.api_version < 500:
+        # Get all volume attachment paths
+        print("\nGet all volume attachment paths for volume attachment at uri: {uri}".format(**volume_attachment_byid))
+        paths = oneview_client.storage_volume_attachments.get_paths(volume_attachment_byid['uri'])
+        for path in paths:
+            print("   Found path at uri: {uri}".format(**path))
 
-    if paths:
-        # Get specific volume attachment path by uri
-        print("Get specific volume attachment path by uri")
-        path_byuri = oneview_client.storage_volume_attachments.get_paths(
-            volume_attachment_byid['uri'], paths[0]['uri'])
-        pprint(path_byuri)
+        if paths:
+            # Get specific volume attachment path by uri
+            print("\nGet specific volume attachment path by uri")
+            path_byuri = oneview_client.storage_volume_attachments.get_paths(
+                volume_attachment_byid['uri'], paths[0]['uri'])
+            pprint(path_byuri)
 
 # Get volume attachment path by id
-try:
-    print("Get volume attachment path by id: '{}'\n  at attachment with id: '{}'".format(path_id, attachment_id))
-    path_byid = oneview_client.storage_volume_attachments.get_paths(attachment_id, path_id)
-    pprint(path_byid)
-except HPOneViewException as e:
-    print(e.msg)
+if path_id and oneview_client.api_version < 500:
+    try:
+        print("\nGet volume attachment path by id: '{}'\n  at attachment with id: '{}'".format(path_id, attachment_id))
+        path_byid = oneview_client.storage_volume_attachments.get_paths(attachment_id, path_id)
+        pprint(path_byid)
+    except HPOneViewException as e:
+        print(e.msg)